all tests passed, norminated, ready to check

+++ philosophers/philo/src/parse_arguments.c
@@ -90,9 +90,9 @@ void	parse_arguments(t_args *args, char **av, int ac)
 			handle_error_str("invalid integer: ", av[i]);
 	}
 	args->num = ft_atoi(av[1]);
-	args->time_to_die = ft_atoi(av[2]);
-	args->time_to_eat = ft_atoi(av[3]);
-	args->time_to_sleep = ft_atoi(av[4]);
+	args->time_to_die = (long long)ft_atoi(av[2]);
+	args->time_to_eat = (long long)ft_atoi(av[3]);
+	args->time_to_sleep = (long long)ft_atoi(av[4]);
 	if (ac == 6)
 		args->must_eat_times = ft_atoi(av[5]);
 	else
